BLENDER: Fix slow patch checking with commits that add or remove many files
Running git update-index for every individual file is slow, so add and remove everything with a single git command.

When such a big commit lands in the default branch, it could cause PR creation and patch checking for all open PRs to be slow, or time out entirely. For example, a commit that removes 1383 files was measured to take more than 60 seconds and timed out. With this change checking took about a second.

This is related to #27967, though this will not help with commits that change many lines in few files.

Co-authored-by: Brecht Van Lommel <brecht@blender.org>