Fix: Slow patch checking with commit that removes many files
Running update-index --remove for every individual file is slow, so
remove them all together.

When such a commit lands in the main branch, it can then cause PR
creation and patch checking for all open PRs to be slow, or time out
entirely.

In our case, a commit that removes 1383 files causes the 60 seconds
timeout to be reached. With this change checking takes about a second.

The git.AddObjectToIndex case most likely needs the same optimization.